# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Connection:
    properties:
      createdAt:
        type: string
      database:
        type: string
      host:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      options:
        additionalProperties: true
        description: Database-specific options (warehouse, role, schema, etc)
        type: object
      password:
        description: AES-256 Encrypted
        type: string
      port:
        type: integer
      type:
        description: postgres, mysql, bigquery, etc
        type: string
      updatedAt:
        type: string
      userId:
        type: string
      username:
        type: string
    type: object
...
  models.Permission:
    properties:
      action:
        description: e.g., 'create', 'read', 'update', 'delete'
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        description: e.g., 'query:create'
        type: string
      resource:
        description: e.g., 'query', 'dashboard'
        type: string
    type: object
  models.QueryExecutionRequest:
    properties:
      connectionId:
        type: string
      limit:
        type: integer
      offset:
        type: integer
      parameters:
        additionalProperties: true
        type: object
      sql:
        type: string
    required:
    - connectionId
    - sql
    type: object
  models.QueryVersion:
    properties:
      aiPrompt:
        type: string
      changeSummary:
        description: e.g., "Modified SQL query"
        type: string
      createdAt:
        type: string
      createdBy:
        description: Metadata
        type: string
      createdByUser:
        $ref: '#/definitions/models.User'
      description:
        type: string
      id:
        type: string
      isAutoSave:
        type: boolean
      metadata:
        description: Additional metadata
        items:
          type: integer
        type: array
      name:
        description: Snapshot data
        type: string
      query:
        allOf:
        - $ref: '#/definitions/models.SavedQuery'
        description: Relationships
      queryId:
        type: string
      sql:
        type: string
      tags:
        items:
          type: integer
        type: array
      version:
        description: Auto-increment per query
        type: integer
      visualizationConfig:
        items:
          type: integer
        type: array
    type: object
  models.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_system_role:
        description: TRUE for built-in roles
        type: boolean
      name:
        type: string
      permissions:
        items:
          $ref: '#/definitions/models.Permission'
        type: array
      updated_at:
        type: string
    type: object
  models.SavedQuery:
    properties:
      aiPrompt:
        type: string
      businessMetricId:
        type: string
      collectionId:
        type: string
      connection:
        allOf:
        - $ref: '#/definitions/models.Connection'
        description: Relationships (optional for queries)
      connectionId:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      pinned:
        type: boolean
      sql:
        type: string
      tags:
        items:
          type: string
        type: array
      updatedAt:
        type: string
      userId:
        type: string
      versions:
        items:
          $ref: '#/definitions/models.QueryVersion'
        type: array
      visualizationConfig:
        description: JSON
        items:
          type: integer
        type: array
    type: object
...
  models.User:
    properties:
      createdAt:
        type: string
      deactivatedAt:
        type: string
      deactivatedBy:
        type: string
      deactivationReason:
        type: string
      email:
        type: string
      emailVerified:
        type: boolean
      emailVerifiedAt:
        type: string
      id:
        type: string
      name:
        type: string
      provider:
        description: OAuth fields
        type: string
      providerId:
        description: OAuth provider user ID
        type: string
      role:
        description: user, admin
        type: string
      roles:
        description: RBAC
        items:
          $ref: '#/definitions/models.Role'
        type: array
      status:
        description: Status
        type: string
      updatedAt:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@insightengine.ai
    name: API Support
  description: Enterprise BI Platform API with Power BI/Tableau parity.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: InsightEngine API
  version: "1.0"
paths:
...
  /query:
    get:
      consumes:
      - application/json
      description: Returns a list of saved queries for the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List saved queries
      tags:
      - Query
    post:
      consumes:
      - application/json
      description: Creates a new saved query.
      parameters:
      - description: Saved Query
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SavedQuery'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a saved query
      tags:
      - Query
  /query/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a saved query by ID.
      parameters:
      - description: Query ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a saved query
      tags:
      - Query
    get:
      consumes:
      - application/json
      description: Returns a single saved query by ID.
      parameters:
      - description: Query ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get a saved query
      tags:
      - Query
    put:
      consumes:
      - application/json
      description: Updates an existing saved query.
      parameters:
      - description: Query ID
        in: path
        name: id
        required: true
        type: string
      - description: Saved Query Updates
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SavedQuery'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a saved query
      tags:
      - Query
  /query/{id}/run:
    post:
      consumes:
      - application/json
      description: Executes a saved query by ID with optional limit and offset.
      parameters:
      - description: Query ID
        in: path
        name: id
        required: true
        type: string
      - description: Run Parameters (limit, offset)
        in: body
        name: request
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Run a saved query
      tags:
      - Query
  /query/execute:
    post:
      consumes:
      - application/json
      description: Executes a raw SQL query on a specific connection.
      parameters:
      - description: Query Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.QueryExecutionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Execute ad-hoc query
      tags:
      - Query
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"